pyproject.toml
[tool.poetry]
name = "first"
version = "0.1.0"
description = ""
authors = ["Павел <stalk200213@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
ipython = "^8.19.0"
python-dotenv = "^1.0.0"
django-taggit = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
